requirements.txt
flask
flask-cors
torch
torchvision
opencv-python
numpy
pillow
matplotlib
pip-system-certs
